Revert "I don't pretend to know why these ampersands are there, but adding the missing initial one fixes things for me"
This reverts commit a5bc1cf8f0181462791e36a80de96a54e5a94bd6.

The ampersands are part of the whacky filename that causes things to break, but
only on very few platforms. If this test fails for you too now, congrats, you
are the first developer to reproduce this and have the honor to fix the problem
(It's something about basename() being locale dependent. ;-)