Bugfix Vfs Smb driver top level dirs are not correctly parsed When upgrading a system from openSUSE Leap 15.2 to Leap 15.3, Gollem starts giving garbled output for top level dir when using Vfs Smb Backend Seems like either PHP or the PCRE library has a BC break here.
This bug fix is a conservative approach at modernizing the parsing regex.
Seems like quantification (+) of a literal space character no longer works as expected
Replacing them with the symbolic whitespace class with quantifier \s+ seems to fix this.
Added some comments for readability.

This should be covered by a test suite but no time to build one right now.